renamed some files, impose 1000 simultaneous game limit
