+++ lib/tictactoe/tttHelper.js
@@ -1,5 +1,7 @@
 const { MessageActionRow, MessageButton } = require('discord.js');
-const { game } = require('./ttcGame');
+const { game } = require('./tttGame');
+
+const maxNumGames = 1000;
 
 var gamesAlive = new Object();
 
@@ -38,6 +40,12 @@ async function createGame(interaction) {
     if (!interaction.isCommand()) return;
     const newGame = new game();
     const newGameId = createGameId();
+
+    // limit on number of games
+    if (Object.keys(gamesAlive).length >= maxNumGames) {
+        delete gamesAlive[Object.keys(gamesAlive)[0]];
+    }
+
     gamesAlive[newGameId] = newGame;
 
     interaction.reply({
